Declaring, casting to unsigned int in cython code
The Cython docs aren't clear about this, but they suggest casting integers to unsigned
when possible so that Cython doesn't bounds check. The docs also suggest that at some point
in the future sign information may be gleaned from the declared numpy dtype, but explicitly
casting to unsigned comes with no performance hit (or improvement) so I might as well do it.